Extract cancellable fetch
